Moved theme switching functionality  to account page
